fix:修改mobx模板文件注入形式，将顶层定义类型改为import导入，优化方法获取preset的方式
